# backend/src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: ParDespue API
  description: API para gestionar usuarios, roles, cápsulas, contenidos, notificaciones y más.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Servidor local
paths:
  /users:
    get:
      summary: Obtener todos los usuarios
      tags: [Users]
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Error interno del servidor
    post:
      summary: Crear un nuevo usuario
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Error de validación
        '500':
          description: Error interno del servidor
  /users/{id}:
    get:
      summary: Obtener un usuario por ID
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
          description: ID del usuario
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Usuario no encontrado
        '500':
          description: Error interno del servidor
    put:
      summary: Actualizar un usuario
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '200':
          description: Usuario actualizado exitosamente
        '400':
          description: Error de validación
        '404':
          description: Usuario no encontrado
        '500':
          description: Error interno del servidor
    delete:
      summary: Eliminar un usuario
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Usuario eliminado exitosamente
        '404':
          description: Usuario no encontrado
        '500':
          description: Error interno del servidor
  /users/{id}/password:
    put:
      summary: Cambiar contraseña de usuario
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword: { type: string }
                newPassword: { type: string }
      responses:
        '200':
          description: Contraseña actualizada correctamente
        '400':
          description: Error de validación
        '403':
          description: No autorizado
        '404':
          description: Usuario no encontrado
        '500':
          description: Error interno del servidor
  /users/{id}/profile-picture:
    put:
      summary: Actualizar foto de perfil
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                profile_picture:
                  type: string
                  format: binary
      responses:
        '200':
          description: Foto de perfil actualizada
        '400':
          description: No se subió ninguna imagen
        '500':
          description: Error interno del servidor
  /users/me/export:
    post:
      summary: Exportar todos los datos del usuario autenticado
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password: { type: string }
      responses:
        '200':
          description: Datos exportados correctamente
        '401':
          description: Contraseña incorrecta
        '500':
          description: Error interno del servidor
...
  /upload:
    post:
      summary: Subir archivo definitivo a una cápsula
      tags: [Upload]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                userId:
                  type: integer
                capsuleId:
                  type: integer
      responses:
        '200':
          description: Archivo subido correctamente
        '400':
          description: No se subió ningún archivo
        '500':
          description: Error interno del servidor
  /upload/tmp:
    post:
      summary: Subir archivo temporal (antes de asociar a cápsula)
      tags: [Upload]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                userId:
                  type: integer
      responses:
        '200':
          description: Archivo temporal subido correctamente
        '400':
          description: No se subió ningún archivo
        '500':
          description: Error interno del servidor
  /upload/move:
    post:
      summary: Mover archivo temporal a carpeta definitiva de cápsula
      tags: [Upload]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId: { type: integer }
                capsuleId: { type: integer }
                tmpPath: { type: string }
      responses:
        '200':
          description: Archivo movido correctamente
        '400':
          description: Faltan datos requeridos
        '500':
          description: Error interno del servidor
  /upload/delete:
    delete:
      summary: Eliminar archivo físico del sistema de archivos
      tags: [Upload]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filePath: { type: string }
      responses:
        '200':
          description: Archivo eliminado
        '400':
          description: Falta filePath
        '500':
          description: Error interno del servidor
components:
  schemas:
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        Profile_Picture: { type: string }
    UserInput:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
...